Move package to suggestion.
